avoid JIT in write
